refactor(icon-button): refactor component with styles abstraction
Refactored the IconButton component to use a centralized styles abstraction for improved maintainability and readability. Introduced `button.style.ts` and `button.interface.ts` to define reusable and type-safe style configurations. Removed redundant class handling logic within the component for a more streamlined implementation.
